requirements.txt
Django>=4.2
psycopg2-binary>=2.9
django-redis>=6.0
python-dotenv>=1.0
Pillow
django-filter
